Finished another round of tests:
- The Board:
The prototype board soldered by Zalman works well.
Still need to solder risers, and push the ESP onto the board.

- The Sensor File:
The sensor python file needs to multithread, and designed better.
Additionally, the *transmission* of -
the data_dict to server, and the *motion detection* DO NOT WORK.

* I added a backup file for a version of sensor.py which looks better-
organized, but also doesn't work.